refactor(invoices): extract invoice mutation lifecycle slice
