+++ backend/app/features/invoices/mutation/__init__.py
@@ -1,0 +1,5 @@
+"""Invoice mutation lifecycle slice."""
+
+from app.features.invoices.mutation.service import InvoiceMutationService
+
+__all__ = ["InvoiceMutationService"]
